fix: 엣지 케이스 수정 (계산기 소수점 대응, 0으로 나누기 대응)

+++ src/main/kotlin/step2/domain/Calculator.kt
@@ -2,7 +2,7 @@ package step2.domain
 
 object Calculator {
 
-    fun calculate(calculatorInput: CalculatorInput): Long {
+    fun calculate(calculatorInput: CalculatorInput): Double {
         val operators = calculatorInput.operators
         val numbers = calculatorInput.numbers
